examples/pyproject.toml
[project]
name = "examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "authlib==1.6.0",
    "fastapi==0.115.12",
    "httpx==0.28.1",
    "python-dotenv==1.1.0",
    "requests==2.32.2",
    "uvicorn==0.34.2",
]
